server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const AuthorizationError = require('../../exceptions/AuthorizationError')
 const { mapDBToModel } = require('../../utils')
 
 class NotesService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._pool = new Pool()
     this._collaborationsService = collaborationsService
+    this._cacheService = cacheService
   }
 
   async addNote({ title, tags, body, owner }) {
@@ -24,19 +25,31 @@ class NotesService {
     if (!result.rows[0].id) {
       throw new InvariantError('Catatan gagal ditambahkan')
     }
+    await this._cacheService.delete(`notes:${owner}`)
     return result.rows[0].id
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-    LEFT JOIN collaborations ON collaborations.note_id = notes.id
-    WHERE notes.owner = $1 OR collaborations.user_id = $1
-    GROUP BY notes.id`,
-      values: [owner]
+    // mendapatkan catatan dari cache
+    try {
+      const result = await this._cacheService.get(`notes:${owner}`)
+      return JSON.parse(result)
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+      LEFT JOIN collaborations ON collaborations.note_id = notes.id
+      WHERE notes.owner = $1 OR collaborations.user_id = $1
+      GROUP BY notes.id`,
+        values: [owner]
+      }
+      const result = await this._pool.query(query)
+      const mappedResult = result.rows.map(mapDBToModel)
+
+      // catatan akan disimpan pada cache sebelum fungsi getNotes dikembalikan
+      await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult))
+
+      return mappedResult
     }
-    const result = await this._pool.query(query)
-    return result.rows.map(mapDBToModel)
   }
 
   async getNoteById(id) {
@@ -57,24 +70,30 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString()
     const query = {
-      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id',
+      text: 'UPDATE notes SET title=$1, body=$2, tags=$3, updated_at=$4 WHERE id=$5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id]
     }
     const result = await this._pool.query(query)
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    await this._cacheService.delete(`notes:${owner}`)
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id]
     }
     const result = await this._pool.query(query)
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    await this._cacheService.delete(`notes:${owner}`)
   }
 
   async verifyNoteOwner(id, owner) {
